Synthetic commit for incomplete tag OPENBSD_3_7_BASE

+++ gnu/egcs/gcc/fixinc/inclhack.tpl
@@ -118,7 +118,12 @@ for INPUT in ${INPUTLIST} ; do
 
 cd ${ORIGDIR}
 
-cd ${INPUT} || continue
+# This originally used cd || continue, however, that does not work with the
+# Solaris2 /bin/sh.
+if [ ! -d ${INPUT} ]; then
+  continue
+fi
+cd ${INPUT}
 INPUT=`${PWDCMD}`
 
 #
@@ -403,9 +408,9 @@ find . -name DONE -exec rm -f '{}' ';'
 
 echo 'Removing unneeded directories:'
 cd $LIB
-all_dirs=`find . -type d -print | sort -r`
+all_dirs=`find . -type d \! -name '.' -print | sort -r`
 for file in $all_dirs; do
-  rmdir $LIB/$file > /dev/null 2>&1 | :
+  rmdir $LIB/$file > /dev/null 2>&1
 done
 
 # # # # # # # # # # # # # # # # # # # # #
@@ -416,11 +421,13 @@ done
 #
 # # # # # # # # # # # # # # # # # # # # #
 
-cd $ORIGDIR
-rm -f include/assert.h
-cp ${srcdir}/assert.h include/assert.h || exit 1
-chmod a+r include/assert.h
-[=
+if [ x${INSTALL_ASSERT_H} != x ] && [ -f ${srcdir}/assert.h ]
+then
+  cd $ORIGDIR
+  rm -f include/assert.h
+  cp ${srcdir}/assert.h include/assert.h || exit 1
+  chmod a+r include/assert.h
+fi[=
 
 #  Make the output file executable
 # =][=
